+++ apps/cowswap-frontend/src/modules/tradeFlow/hooks/useSolanaTradeFlowContext.ts
@@ -26,16 +26,6 @@ import { getSolanaDelegationAmount } from '../utils/getSolanaDelegationAmount'
 import { getSolanaSellToken } from '../utils/getSolanaSellToken'
 import { getUiOrderType } from '../utils/getUiOrderType'
 
-// Mirrors swapFlow's `orderParams.recipient = recipientAddress || recipient || account`: prefer the
-// resolved recipient address, then the raw recipient value, then default to sending to self.
-export function resolveSolanaReceiver(params: {
-  recipient: string | null | undefined
-  recipientAddress: string | null | undefined
-  account: string
-}): string {
-  return params.recipientAddress || params.recipient || params.account
-}
-
 export function useSolanaTradeFlowContext({ deadline }: TradeFlowParams): SolanaTradeFlowContext | null {
   const { chainId, account } = useWalletInfo()
   const derivedTradeState = useDerivedTradeState()
