+++ apps/cowswap-frontend/src/modules/tradeFlow/utils/getIsSolanaTradeFlowContextReady.test.ts
@@ -3,7 +3,7 @@ import { CurrencyAmount, Token } from '@cowprotocol/currency'
 import { UiOrderType } from '@cowprotocol/types'
 
 import { getIsSolanaTradeFlowContextReady } from './getIsSolanaTradeFlowContextReady'
-import { resolveSolanaReceiver } from './useSolanaTradeFlowContext'
+import { resolveSolanaReceiver } from './resolveSolanaReceiver'
 
 // The canonical Solana System Program address (32 zero bytes) — always a valid
 // Solana pubkey, used here purely as "some syntactically valid Solana address".
